src/python/vega/features/core/serializers/generation_request_serializer.py

from rest_framework import serializers

from features.core.models import GenerationRequest


class GenerationRequestSerializer(serializers.ModelSerializer):

	class Meta:
		model = GenerationRequest
		fields = ["song", "status"]
