requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
audioowl==0.0.14
audioread==3.0.0
backcall==0.2.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.2
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
dropbox==11.27.0
executing==1.2.0
fonttools==4.43.1
frozenlist==1.4.0
idna==3.4
ipython==8.14.0
jedi==0.19.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.40.1
madmom==0.15.1
Markdown==3.4.4
matplotlib==3.8.0
matplotlib-inline==0.1.6
mixingbear==0.0.1
msgpack==1.0.7
multidict==6.0.4
noisereduce==3.0.0
numba==0.57.1
numpy==1.24.4
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
ply==3.11
pooch==1.7.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydub==0.20.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
resampy==0.4.2
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
stack-data==0.6.2
stone==3.3.1
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.8.0
urllib3==2.0.4
wcwidth==0.2.6
yarl==1.9.2
